update configuration files as per review comments
